# Track CoW pages to avoid iterating over the page tables