--- a/lib/tinykvm/memory.cpp
+++ b/lib/tinykvm/memory.cpp
@@ -1,4 +1,5 @@
 #include "machine.hpp"
+#include <algorithm>
 #include <cstring>
 #include <sys/mman.h>
 #include <stdexcept>
@@ -53,7 +54,8 @@ void vMemory::record_cow_page(uint64_t addr, uint64_t entry)
 	// If the page is writable, we will restore the original
 	// memory from the master VM. We only care about leaf pages.
 	if (machine.is_forked() && entry & PDE64_USER) {
-		cow_written_pages.push_back(addr);
+		auto it = std::lower_bound(cow_written_pages.begin(), cow_written_pages.end(), addr);
+		cow_written_pages.insert(it, addr);
 	}
 }