--- a/lib/tinykvm/memory.hpp
+++ b/lib/tinykvm/memory.hpp
@@ -4,7 +4,6 @@
 #include "virtual_mem.hpp"
 #include <cstddef>
 #include <mutex>
-#include <set>
 #include <string_view>
 
 namespace tinykvm {
@@ -17,7 +16,7 @@ struct vMemory {
 	}
 
 	Machine& machine;
-	std::set<uint64_t> cow_written_pages{};
+	std::vector<uint64_t> cow_written_pages{};
 	uint64_t physbase;
 	uint64_t safebase;
 	uint64_t page_tables;