Most of the time spent in fork_reset is spent iterating over the page tables. For kvmserver with threads=1 I see a reduction from 280us to 135us for deno hello world and 900us to 735us for the Deno React benchmark.

The existing implementation always picked up page 0xffe00000 which had special flags. This is unrelated to this change but will remove since it appears to be unnecessary. See: https://github.com/varnish/tinykvm/blob/6af8ab6aeb494db045ed8e6516670bbc588c205f/lib/tinykvm/amd64/paging.cpp#L529

The MADV_FREE is also changed to MADV_DONTNEED which guarantees the page will be zeroed on subsequent access.