@lrowe 2025-06-14
This is very possibly a misunderstanding on my part but I am unsure if the MADV_FREE logic in fork_reset is correct.

While the logic seems to hold if the page is actually freed by the kernel (which would effectively zero it), that is not guaranteed to happen immediately. If the page is written to again before that happens the madvise man page suggests we would see the current data in the next request.